pyproject.toml
[project]
name = "llm-council"
version = "1.1.0"
description = "LLM Council - Multi-Provider AI Council System"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "tenacity>=9.0.0",
    "slowapi>=0.1.9",
    "motor>=3.7.0",
    "pymongo>=4.15.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
]
